add static_cast<int> in two instances...
...of taking the negative of a unsigned number which may not work as
expected in some cases (and msvc complains with a warning)

+++ src/gui/widgets/text.cpp
@@ -220,7 +220,7 @@ void ttext_::set_cursor(const size_t offset, const bool select)
 		if(selection_start_ == offset) {
 			selection_length_ = 0;
 		} else {
-			selection_length_ = - (selection_start_ - offset);
+			selection_length_ = - static_cast<int>(selection_start_ - offset);
 		}
 
 #ifdef __unix__
